pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paddleocr-guide"
version = "0.2.0"
description = "PaddleOCR 3.x 实战指南 - 中文 OCR 最佳实践 | PaddleOCR 3.x Guide - Chinese OCR Best Practices"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Stephen Chan", email = "stephenkaylonchan@gmail.com"}
]
maintainers = [
    {name = "Stephen Chan", email = "stephenkaylonchan@gmail.com"}
]
keywords = [
    "ocr",
    "paddleocr",
    "chinese-ocr",
    "document-parsing",
    "table-recognition",
    "layout-analysis",
    "formula-recognition",
    "paddlepaddle",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Natural Language :: Chinese (Simplified)",
    "Natural Language :: English",
    "Typing :: Typed",
]

dependencies = [
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0",
    "Pillow>=9.0.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "opencv-python>=4.5.0",
    "pandas>=1.3.0",
    "openpyxl>=3.0.0",
    "matplotlib>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/stephenkaylonchan/paddleocr-guide"
Documentation = "https://github.com/stephenkaylonchan/paddleocr-guide#readme"
Repository = "https://github.com/stephenkaylonchan/paddleocr-guide.git"
Issues = "https://github.com/stephenkaylonchan/paddleocr-guide/issues"
Changelog = "https://github.com/stephenkaylonchan/paddleocr-guide/blob/main/CHANGELOG.md"

# =============================================================================
# Setuptools 配置
# =============================================================================

[tool.setuptools.packages.find]
where = ["."]
include = ["examples*"]
exclude = ["tests*", "docs*", ".github*"]

[tool.setuptools.package-data]
"examples" = ["*.md", "*.txt"]

# =============================================================================
# Black 代码格式化
# =============================================================================

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | \.ocr_python_venv
    | __pycache__
    | build
    | dist
    | \.eggs
)/
'''

# =============================================================================
# isort 导入排序
# =============================================================================

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["examples"]
known_third_party = ["paddleocr", "paddlepaddle", "PIL", "numpy", "pandas"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = [".venv/*", ".ocr_python_venv/*", "build/*", "dist/*"]

# =============================================================================
# Pytest 测试配置
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_model: marks tests that require downloading models",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# Coverage 覆盖率配置
# =============================================================================

[tool.coverage.run]
source = ["examples"]
branch = true
omit = [
    "tests/*",
    "examples/_common/logging_config.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# =============================================================================
# MyPy 类型检查
# =============================================================================

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
no_implicit_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "paddleocr.*",
    "paddlepaddle.*",
    "cv2.*",
    "pandas.*",
]
ignore_missing_imports = true

# =============================================================================
# Bandit 安全检查
# =============================================================================

[tool.bandit]
exclude_dirs = ["tests", ".venv", ".ocr_python_venv"]
skips = ["B101"]  # 允许使用 assert 语句

# =============================================================================
# Ruff (可选的快速 linter)
# =============================================================================

[tool.ruff]
target-version = "py38"
line-length = 100
exclude = [
    ".git",
    ".venv",
    ".ocr_python_venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
